refactor: Color, Type enum 분리
