+++ src/main/java/softeer2nd/chess/pieces/Color.java
@@ -1,0 +1,6 @@
+package softeer2nd.chess.pieces;
+
+public enum Color {
+    // 반복문 돌릴시 NO_COLOR 문제
+    WHITE, BLACK, NO_COLOR;
+}
